surgical_system/py_src/requirements.txt
numpy
scipy.spatial.transform